feat(ProductAverageRating): hide review comments based on feature flag
