+++ apps/store/.env.local.example
@@ -49,6 +49,7 @@ NEXT_PUBLIC_FEATURE_WIDGET_SWITCH=<true>
 NEXT_PUBLIC_FEATURE_PRICE_INTENT_WARNING=<true>
 NEXT_PUBLIC_FEATURE_BANKID_V6=<true>
 NEXT_PUBLIC_CROSS_SELL_CARD_V2=<true>
+NEXT_PUBLIC_HIDE_REVIEWS_FROM_PRODUCT_AVERAGE_RATING=<true>
 
 
 # Vercel Edge Config Connection String (https://github.com/vercel/edge-config)
